fix: prevent integer overflow in BFGS lwork calculation

+++ source/source_relax/bfgs.cpp
@@ -128,7 +128,7 @@ void BFGS::PrepareStep(std::vector<ModuleBase::Vector3<double>>& force,
     //! call dysev
     std::vector<double> omega(3*size);
     std::vector<double> work(3*size*3*size);
-    int lwork=3*size*3*size;
+    long long lwork=3*size*3*size;
     int info=0;
     std::vector<double> H_flat;
